% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer.R
\name{analyzer}
\alias{analyzer}
\title{Analyze stuff}
\usage{
analyzer(
  SNPcomparison,
  minDepthPercentile = 0.1,
  maxDepthPercentile = 0.9,
  windowSize = 25000,
  bulkSize = 20,
  outlierFilt = "deltaSNP",
  filter_chr_list = NULL
)
}
\arguments{
\item{minDepthPercentile}{sample depth - Half the min depth, Default: 0.9}

\item{maxDepthPercentile}{depth: higher fifth percentile >
quantile(sum of depth in both bulks, na.rm = T, probs = 0.95)), Default: 0.1}

\item{windowSize}{PARAM_DESCRIPTION, Default: 25000}

\item{bulkSize}{PARAM_DESCRIPTION, Default: 20}

\item{outlierFilt}{PARAM_DESCRIPTION, Default: 'deltaSNP'}

\item{filter_chr_list}{PARAM_DESCRIPTION, Default: NULL}

\item{SNPcomparison:}{lower tenth percentile >
quantile(sum of depth in both bulks, na.rm = T, probs = 0.1))}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Now that we have a list with all comparisons for replicates
(object: "replicates"), pools (object: "pools") and all
samples together (allPoolsInOneComparison) we can begin filtering the
samples. First idea is to filter by the following params
}
\details{
DETAILS
}
