% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binDistPlotter.R
\name{binDistPlotter}
\alias{binDistPlotter}
\title{plot bin dist}
\usage{
binDistPlotter(
  inputBinTable,
  windowsSizes = c(5000, 10000, 20000, 50000),
  fileName
)
}
\arguments{
\item{inputBinTable}{PARAM_DESCRIPTION}

\item{windowsSizes}{PARAM_DESCRIPTION, Default: c(5000, 10000, 20000, 50000)}

\item{fileName}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
This function makes the distribution plots for the bins, using the output table of the binner()palette() function as input.
}
\details{
DETAILS
}
