% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binner.R
\name{binner}
\alias{binner}
\title{Binner -- the first}
\usage{
binner(input, bin.size = 10000)
}
\arguments{
\item{input}{PARAM_DESCRIPTION}

\item{bin.size}{PARAM_DESCRIPTION, Default: 10000}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
This is a function to make a new table with different bin sizes
and the average of ∆\% observed in the positions inside the bin.
You can play with the bin size, and it uses the output of percenter as input.
}
\details{
DETAILS
}
