% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binner2.R
\name{binner2}
\alias{binner2}
\title{Binner the second}
\usage{
binner2(input, bin.size = 10000)
}
\arguments{
\item{input}{PARAM_DESCRIPTION}

\item{bin.size}{PARAM_DESCRIPTION, Default: 10000}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
how diff than binner?

This is a function to make a new table with different
bin sizes and the average of ∆\% observed in the positions inside the bin.
You can play with the bin size, and it uses the output of percenter
as input.
}
\details{
DETAILS
}
\note{
a bin is considered significant if half or more of the
snps in the bin are significant
}
