% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binner3.R
\name{binner3}
\alias{binner3}
\title{Another binner}
\usage{
binner3(input, bin.size = 10000)
}
\arguments{
\item{input}{PARAM_DESCRIPTION}

\item{bin.size}{PARAM_DESCRIPTION, Default: 10000}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
see binner and binner2
}
\details{
DETAILS
}
