% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{comparisons}
\alias{comparisons}
\alias{comparisons<-}
\alias{createComparisonsFrame}
\alias{comparisons,BSAExperiment-method}
\alias{comparisons<-,BSAExperiment-method}
\title{Get the BSAExperiment comparisons frame}
\usage{
comparisons(object, comparisons)

comparisons(object, ...) <- value

createComparisonsFrame(
  object,
  grouping_variable = "batch",
  bulk_variable = "condition",
  bulk_base_condition = "inoculum",
  var1_name = "low_bulk",
  var2_name = "high_bulk",
  base_cond_in_each_group = TRUE
)

\S4method{comparisons}{BSAExperiment}(object)

\S4method{comparisons}{BSAExperiment}(object) <- value
}
\description{
Get the BSAExperiment comparisons frame

Set the BSAExperiment comparison frame
}
