% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_bsa_data.R
\name{collapse_bsa_data}
\alias{collapse_bsa_data}
\title{Collapse BSA (Bulk Segregant Analysis) Data}
\usage{
collapse_bsa_data(df_with_meta, grouping_conditions, sample_columns)
}
\arguments{
\item{df_with_meta}{A dataframe that contains BSA data.}

\item{grouping_conditions}{A vector of column names that specifies the
grouping conditions.}

\item{sample_columns}{A vector of column names that specifies the
sample columns.}
}
\value{
A dataframe in which the data has been grouped, summarized,
united, and arranged according to the specified conditions and columns.
}
\description{
This function takes a dataframe of BSA data, a list of grouping conditions,
and a list of sample columns, and returns a dataframe in which the data has
been grouped according to the specified conditions. It performs several
operations including grouping, summarizing, uniting, and arranging.
}
\examples{
df_with_meta <- data.frame(
  A = rep(c("group1", "group2"), each = 5),
  CHR = rep("CHR1", 10),
  POS = rep(c(1,5,10,20,30), times=2),
  REF_Allele = rep('A', 10),
  ALT1 = rep('G', 10),
  RealDepth = runif(10, 1, 10),
  Depth = runif(10, 1, 10),
  Reference = runif(10, 1, 10),
  Alternative1 = runif(10, 1, 10)
)
grouping_conditions <- c('CHR','POS','REF_Allele','ALT1')
sample_columns <- c("RealDepth", "Depth", "Reference", "Alternative1")

collapse_bsa_data(df_with_meta, grouping_conditions, sample_columns)

}
