% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getG.R
\name{getG}
\alias{getG}
\title{get G prime stat}
\usage{
getG(LowRef, HighRef, LowAlt, HighAlt)
}
\arguments{
\item{LowRef}{PARAM_DESCRIPTION}

\item{HighRef}{PARAM_DESCRIPTION}

\item{LowAlt}{PARAM_DESCRIPTION}

\item{HighAlt}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
