% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPvals.R
\name{getPvals}
\alias{getPvals}
\title{get g prime pvals}
\usage{
getPvals(
  Gprime,
  deltaSNP = NULL,
  outlierFilter = c("deltaSNP", "Hampel"),
  filterThreshold
)
}
\arguments{
\item{Gprime}{PARAM_DESCRIPTION}

\item{deltaSNP}{PARAM_DESCRIPTION, Default: NULL}

\item{outlierFilter}{PARAM_DESCRIPTION, Default: c("deltaSNP", "Hampel")}

\item{filterThreshold}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
