% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percenter2.R
\name{percenter2}
\alias{percenter2}
\title{Another percenter function}
\usage{
percenter2(input, parameter = "Reference", sampleColumn = 1)
}
\arguments{
\item{input}{PARAM_DESCRIPTION}

\item{parameter}{PARAM_DESCRIPTION, Default: 'Reference'}

\item{sampleColumn}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
