% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMakerBSA.R
\name{plotMakerBSA}
\alias{plotMakerBSA}
\title{plot maker BSA}
\usage{
plotMakerBSA(datum, xplot = datum$binMiddle/1000, yplot = datum$All_pools)
}
\arguments{
\item{depth}{integer. A read depth for which to replicate SNP-index calls.}

\item{altFreq1}{numeric. The alternate allele frequency for bulk A.}

\item{altFreq2}{numeric. The alternate allele frequency for bulk B.}

\item{replicates}{integer. The number of bootstrap replications.}

\item{filter}{numeric. an optional minimum SNP-index filter}
}
\value{
Returns a vector of length replicates delta SNP-indices
}
\description{
Identifies which variants are statistically significant in
x number of pools.
}
