% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{plotter}
\alias{plotter}
\title{plot some stuff}
\usage{
plotter(input, fileName)
}
\arguments{
\item{input}{PARAM_DESCRIPTION}

\item{fileName}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
This function will plot the enrichment of each SNP.
It uses the output of the percenter() as input.
Obs, I have to set to run the percenter() as well as define the
fileNames outside the function, in the for loop to make it work.
here what I mean:  l=percenter(sampleColumn=sampleColumn)
}
\details{
DETAILS
}
