% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGprimeAnalysis.R
\name{runGprimeAnalysis}
\alias{runGprimeAnalysis}
\title{analyze the g prime}
\usage{
runGprimeAnalysis(
  SNPset,
  windowSize = 1e+06,
  outlierFilter = "deltaSNP",
  filterThreshold = 0.1,
  ...
)
}
\arguments{
\item{SNPset}{PARAM_DESCRIPTION}

\item{windowSize}{PARAM_DESCRIPTION, Default: 1e+06}

\item{outlierFilter}{PARAM_DESCRIPTION, Default: 'deltaSNP'}

\item{filterThreshold}{PARAM_DESCRIPTION, Default: 0.1}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
