% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQTLseqAnalysis.R
\name{runQTLseqAnalysis_local}
\alias{runQTLseqAnalysis_local}
\title{run QTL seq analysis}
\usage{
runQTLseqAnalysis_local(
  SNPset,
  windowSize = 1e+06,
  popStruc = "F2",
  bulkSize,
  depth = NULL,
  replications = 10000,
  filter = 0.3,
  intervals = c(95, 99)
)
}
\arguments{
\item{SNPset}{PARAM_DESCRIPTION}

\item{windowSize}{PARAM_DESCRIPTION, Default: 1e+06}

\item{popStruc}{PARAM_DESCRIPTION, Default: 'F2'}

\item{bulkSize}{PARAM_DESCRIPTION}

\item{depth}{PARAM_DESCRIPTION, Default: NULL}

\item{replications}{PARAM_DESCRIPTION, Default: 10000}

\item{filter}{PARAM_DESCRIPTION, Default: 0.3}

\item{intervals}{PARAM_DESCRIPTION, Default: c(95, 99)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
