% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateConfInt.R
\name{simulateConfInt}
\alias{simulateConfInt}
\title{a simulate function}
\usage{
simulateConfInt(
  SNPset,
  popStruc = "F2",
  bulkSize,
  depth = 1:100,
  replications = 10000,
  filter = 0.3,
  intervals = c(0.05, 0.025)
)
}
\arguments{
\item{SNPset}{PARAM_DESCRIPTION}

\item{popStruc}{PARAM_DESCRIPTION, Default: 'F2'}

\item{bulkSize}{PARAM_DESCRIPTION}

\item{depth}{PARAM_DESCRIPTION, Default: 1:100}

\item{replications}{PARAM_DESCRIPTION, Default: 10000}

\item{filter}{PARAM_DESCRIPTION, Default: 0.3}

\item{intervals}{PARAM_DESCRIPTION, Default: c(0.05, 0.025)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
