% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateSNPindex.R
\name{simulateSNPindex_local}
\alias{simulateSNPindex_local}
\title{simulate snp index}
\usage{
simulateSNPindex_local(
  depth,
  altFreq1,
  altFreq2,
  replicates = 10000,
  filter = NULL
)
}
\arguments{
\item{depth}{PARAM_DESCRIPTION}

\item{altFreq1}{PARAM_DESCRIPTION}

\item{altFreq2}{PARAM_DESCRIPTION}

\item{replicates}{PARAM_DESCRIPTION, Default: 10000}

\item{filter}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
