% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{tabler}
\alias{tabler}
\title{the tabler}
\usage{
tabler(VCF, i = 1, minDepth = 5, cutoff = 0.75, filtering = F, cutoff50 = 0.1)
}
\arguments{
\item{VCF}{PARAM_DESCRIPTION}

\item{i}{PARAM_DESCRIPTION, Default: 1}

\item{minDepth}{PARAM_DESCRIPTION, Default: 5}

\item{cutoff}{PARAM_DESCRIPTION, Default: 0.75}

\item{filtering}{PARAM_DESCRIPTION, Default: F}

\item{cutoff50}{PARAM_DESCRIPTION, Default: 0.1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
