% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_to_qtlseqr_table.R
\name{tidy_metrics}
\alias{tidy_metrics}
\title{Tidy data from gds file}
\usage{
tidy_metrics(data_mat, values_cname, rnames, cnames)
}
\arguments{
\item{data_mat}{a data matrix in sample x variant format}

\item{values_cname}{name of the metric -- eg RealDepth}

\item{rnames}{vector to name the columns, eg sample.ids}

\item{cnames}{vector to name the rows, eg variant_ids}
}
\value{
a tidy dataframe in long format with columns sample, variant, metric
}
\description{
Tidy data to long format with column sample, variant, and the
metric name. internal use only
}
