% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_variants.R
\name{tile_metrics}
\alias{tile_metrics}
\title{Tile Metrics Dataframe}
\usage{
tile_metrics(
  metrics_df,
  genome_tile_df,
  chr_seqlength,
  cut_col = "POS",
  left_inclusive = TRUE,
  right_inclusive = FALSE
)
}
\arguments{
\item{metrics_df}{PARAM_DESCRIPTION}

\item{genome_tile_df}{PARAM_DESCRIPTION}

\item{chr_seqlength}{PARAM_DESCRIPTION}

\item{cut_col}{PARAM_DESCRIPTION, Default: 'POS'}

\item{left_inclusive}{PARAM_DESCRIPTION, Default: TRUE}

\item{right_inclusive}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}

}
\seealso{
\code{\link[dplyr]{mutate}}
\code{\link[rlang]{sym}}
}
