% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricubeStat.R
\name{tricubeStat}
\alias{tricubeStat}
\title{calc tricube stat}
\usage{
tricubeStat(POS, Stat, windowSize = 2e+06)
}
\arguments{
\item{POS}{PARAM_DESCRIPTION}

\item{Stat}{PARAM_DESCRIPTION}

\item{windowSize}{PARAM_DESCRIPTION, Default: 2e+06}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
