% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binner2.R
\name{binner2}
\alias{binner2}
\title{Binner the second}
\usage{
binner2(input, bin.size = 10000)
}
\arguments{
\item{input}{PARAM_DESCRIPTION}

\item{bin.size}{PARAM_DESCRIPTION, Default: 10000}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Different from binner1
}
\details{
DETAILS
}
