% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percenter.R
\name{percenter}
\alias{percenter}
\title{a percenter function}
\usage{
percenter(input = allSamples, parameter = "Reference", sampleColumn = 1)
}
\arguments{
\item{input}{PARAM_DESCRIPTION, Default: allSamples}

\item{parameter}{PARAM_DESCRIPTION, Default: 'Reference'}

\item{sampleColumn}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
