% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickers.R
\name{picker2}
\alias{picker2}
\title{Transform VCF Table for QTLseqR}
\usage{
picker2(
  samples_with_meta_df,
  low_bulk_sample,
  high_bulk_sample,
  sample_col = "sample",
  bulk_col = "bulk"
)
}
\arguments{
\item{samples_with_meta_df}{A dataframe representing samples
\code{\link[BSA]{vcf_to_qtlseqr_table}} joined with the sample metadata}

\item{low_bulk_sample}{String. The name of the sample (and of the element in
the vcf_df list) corresponding to the Low Bulk.}

\item{high_bulk_sample}{The name of the sample (and of the element in
the vcf_df list) corresponding to the High Bulk.}

\item{sample_col}{name of the column with the sample names. Must contain
the low_bulk_sample and high_bulk_sample. Default 'sample'.}

\item{bulk_col}{name of the column which stores the different sample comparisons.
bulk is a name retained from QTLSeqR and traditional BSA experiments}
}
\value{
a data.frame with the format that can be understood by QTLseq.
}
\description{
Function to set comparisons between two samples and output
the table in a way that the QTLseq package can use. Uses two element of
a list to create a table that can be used by the package QTLseq.
}
\details{
This will be in the format used
by QTLseq package. Picker2 has some adaptations from picker 1 for
the tabler function in awk. RealDepth instead of realDepth,
and no Alt2 allele.
}
