% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickers.R
\name{picker2}
\alias{picker2}
\title{pick something}
\usage{
picker2(SNPset, lowBulk, highBulk)
}
\arguments{
\item{SNPset}{PARAM_DESCRIPTION}

\item{lowBulk}{PARAM_DESCRIPTION}

\item{highBulk}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
