% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickers.R
\name{picker2}
\alias{picker2}
\title{pick something}
\usage{
picker2(SNPset, lowBulk, highBulk)
}
\arguments{
\item{SNPset}{List. Each eelement of the list is a table showing data
from one sample, the table is set as the output of the tabler function.}

\item{lowBulk}{String. The name of the sample (and of the element in
the SNPset list) corresponding to the Low Bulk.}

\item{highBulk}{The name of the sample (and of the element in
the SNPset list) corresponding to the High Bulk.}
}
\value{
a data.frame with the format that can be understood by QTLseq.
}
\description{
Function to set comparisons between two samples and output
the table in a way that the QTLseq package can use. Uses two element of
a list to create a table that can be used by the package QTLseq.
}
\details{
This will be in the format used
by QTLseq package. Picker2 has some adaptations from picker 1 for
the tabler function in awk. RealDepth instead of realDepth,
and no Alt2 allele.
}
